+++ src/main/java/com/ktb/interview/application/service/flow/InterviewSessionFeedbackQueryFlowService.java
@@ -22,8 +22,8 @@ import com.ktb.interview.dto.response.SessionFeedbackFailedResponse;
 import com.ktb.interview.dto.response.SessionFeedbackPendingResponse;
 import com.ktb.interview.exception.InterviewSessionAccessDeniedException;
 import com.ktb.interview.exception.InterviewSessionInvalidStateException;
-import com.ktb.interview.session.persistence.repository.InterviewSessionEntityRepository;
-import com.ktb.interview.session.persistence.repository.projection.InterviewSessionFinalFeedbackReadModel;
+import com.ktb.interview.repository.InterviewSessionEntityRepository;
+import com.ktb.interview.repository.projection.InterviewSessionFinalFeedbackReadModel;
 import com.ktb.interview.session.service.InterviewSessionService;
 import java.util.ArrayList;
 import java.util.List;
