+++ src/main/java/com/ktb/interview/repository/InterviewSessionFeedbackEntityRepository.java
@@ -1,4 +1,4 @@
-package com.ktb.interview.session.persistence.repository;
+package com.ktb.interview.repository;
 
 import com.ktb.interview.domain.entity.InterviewSessionFeedbackEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
