+++ src/main/java/com/ktb/interview/repository/impl/InMemoryInterviewSessionRepository.java
@@ -1,7 +1,7 @@
-package com.ktb.interview.session.repository.impl;
+package com.ktb.interview.repository.impl;
 
 import com.ktb.interview.domain.InterviewSession;
-import com.ktb.interview.session.repository.InterviewSessionRepository;
+import com.ktb.interview.repository.InterviewSessionRepository;
 import java.time.LocalDateTime;
 import java.util.Collection;
 import java.util.Map;
