+++ src/main/java/com/ktb/interview/repository/impl/JpaInterviewSessionFinalFeedbackStore.java
@@ -1,4 +1,4 @@
-package com.ktb.interview.session.persistence.adapter;
+package com.ktb.interview.repository.impl;
 
 import com.ktb.answer.domain.TurnType;
 import com.ktb.interview.dto.ai.InterviewBadCaseFeedbackResponse;
@@ -7,7 +7,7 @@ import com.ktb.interview.dto.ai.InterviewFeedbackMetricResponse;
 import com.ktb.interview.dto.ai.InterviewKeywordResultResponse;
 import com.ktb.interview.dto.ai.InterviewOverallFeedbackResponse;
 import com.ktb.interview.dto.ai.InterviewTopicFeedbackResponse;
-import com.ktb.interview.port.out.InterviewSessionFinalFeedbackStore;
+import com.ktb.interview.repository.InterviewSessionFinalFeedbackStore;
 import com.ktb.interview.domain.InterviewHistoryItem;
 import com.ktb.interview.domain.InterviewQuestionSnapshot;
 import com.ktb.interview.domain.InterviewSession;
@@ -17,11 +17,11 @@ import com.ktb.interview.domain.entity.InterviewSessionFeedbackEntity;
 import com.ktb.interview.domain.entity.InterviewSessionMetricEntity;
 import com.ktb.interview.domain.entity.InterviewTopicFeedbackEntity;
 import com.ktb.interview.domain.entity.InterviewTurnEntity;
-import com.ktb.interview.session.persistence.repository.InterviewSessionEntityRepository;
-import com.ktb.interview.session.persistence.repository.InterviewSessionFeedbackEntityRepository;
-import com.ktb.interview.session.persistence.repository.InterviewSessionMetricEntityRepository;
-import com.ktb.interview.session.persistence.repository.InterviewTopicFeedbackEntityRepository;
-import com.ktb.interview.session.persistence.repository.InterviewTurnEntityRepository;
+import com.ktb.interview.repository.InterviewSessionEntityRepository;
+import com.ktb.interview.repository.InterviewSessionFeedbackEntityRepository;
+import com.ktb.interview.repository.InterviewSessionMetricEntityRepository;
+import com.ktb.interview.repository.InterviewTopicFeedbackEntityRepository;
+import com.ktb.interview.repository.InterviewTurnEntityRepository;
 import com.ktb.metric.domain.Metric;
 import com.ktb.metric.repository.MetricRepository;
 import java.math.BigDecimal;
