+++ src/main/java/com/ktb/interview/session/service/impl/InterviewSessionServiceImpl.java
@@ -7,8 +7,8 @@ import com.ktb.interview.domain.InterviewSession;
 import com.ktb.interview.exception.InterviewSessionAccessDeniedException;
 import com.ktb.interview.exception.InterviewSessionExpiredException;
 import com.ktb.interview.exception.InterviewSessionNotFoundException;
-import com.ktb.interview.session.repository.InterviewSessionFeedbackRepository;
-import com.ktb.interview.session.repository.InterviewSessionRepository;
+import com.ktb.interview.repository.InterviewSessionFeedbackRepository;
+import com.ktb.interview.repository.InterviewSessionRepository;
 import com.ktb.interview.session.service.InterviewSessionService;
 import com.ktb.question.domain.QuestionType;
 import java.time.Duration;
